оновлює файли ДЗ №3
